positions int -> float

+++ sysproduction/interactive_update_roll_status.py
@@ -65,7 +65,7 @@ def get_rolling_master_function():
 class RollDataWithStateReporting(object):
     instrument_code: str
     original_roll_status: RollState
-    position_priced_contract: int
+    position_priced_contract: float # TODO AG position int -> float
     allowable_roll_states_as_list_of_str: list
     days_until_roll: int
     relative_volume: float
@@ -79,7 +79,7 @@ class RollDataWithStateReporting(object):
         print(landing_strip(80))
         print("Current State: %s" % self.original_roll_status)
         print(
-            "Current position in priced contract %d (if zero can Roll Adjusted prices)" %
+            "Current position in priced contract %.2f (if zero can Roll Adjusted prices)" %
             self.position_priced_contract)
         print("")
         print("These are your options:")
@@ -365,9 +365,10 @@ def setup_roll_data_with_state_reporting(data: dataBlob, instrument_code: str) -
 
     contract = futuresContract(instrument_code, priced_contract_date)
 
-    position_priced_contract = int(
-        diag_positions.get_position_for_contract(contract)
-    )
+    # position_priced_contract = int(
+    #     diag_positions.get_position_for_contract(contract)
+    # )
+    position_priced_contract = diag_positions.get_position_for_contract(contract)
 
     allowable_roll_states = allowable_roll_state_from_current_and_position(
         original_roll_status, position_priced_contract
